feat: AA-773: Add in special logic to mark special exams as complete
We have had numerous support tickets/bugs related to special exams
not properly displaying (see https://openedx.atlassian.net/browse/AA-773 for
details). This now looks at the proctoring service to indicate if the exam
is complete and if so, will mark the subsection and all children as complete.

Note: This is purely cosmetic and is not being recorded in the database. This
logic just affects the completion checkmarks.